requirements.txt
fastapi
python-jose
passlib
pymongo
python-multipart
bcrypt
pydantic
hashlib